+++ packages/cli/src/lib/setup/customError.ts
